requirements.txt

torch
opencv-python


ultralytics
Pillow
google-genai